Creating elections with specific ids should update `api_authevent_id_seq` (#273)
Fixes #270

When creating elections with specific ids, the id automatic sequence number (`api_authevent_id_seq`  in postgresql) is not updated. This means that when creating an election with no specific id, it might fail.

When creating elections with specific ids, the the id automatic sequence number (`api_authevent_id_seq`  in postgresql) should be updated to the `max(api_authevent_id_seq, biggest(authevent.id))`.