Race condition in parent-children election between allow-tally and tally
Parent issue: https://github.com/sequentech/meta/issues/115

- Current Behavior

With a set of parent-children elections and having many children (100 for example), if allow-tally is clicked in the parent, then the parent election is allowed before the children. Therefore, tally can be launched when some children are still not tallyable, and this can create errors.

- Expected Behavior

Parent election should be the last to be allowed to tally in parent-children elections: when it is allowed to tally, all its children should have also been allowed to tally already.